+++ python_modules/dagster/dagster/cli/workspace/load.py
@@ -15,6 +15,7 @@ from .autodiscovery import (
     LoadableTarget,
     loadable_targets_from_python_file,
     loadable_targets_from_python_module,
+    loadable_targets_from_python_package,
 )
 from .config_schema import ensure_workspace_config
 from .workspace import Workspace
@@ -55,8 +56,12 @@ def load_def_in_module(module_name, attribute):
     return def_from_pointer(CodePointer.from_module(module_name, attribute))
 
 
-def load_def_in_python_file(python_file, attribute):
-    return def_from_pointer(CodePointer.from_python_file(python_file, attribute))
+def load_def_in_package(package_name, attribute):
+    return def_from_pointer(CodePointer.from_python_package(package_name, attribute))
+
+
+def load_def_in_python_file(python_file, attribute, working_directory):
+    return def_from_pointer(CodePointer.from_python_file(python_file, attribute, working_directory))
 
 
 def _location_handle_from_module_config(python_module_config):
@@ -100,6 +105,47 @@ def location_handle_from_module_name(module_name, attribute, location_name=None)
     )
 
 
+def _location_handle_from_package_config(python_module_config):
+    module_name, attribute, location_name = _get_package_config_data(python_module_config)
+    return location_handle_from_package_name(module_name, attribute, location_name)
+
+
+def _get_package_config_data(python_package_config):
+    return (
+        (python_package_config, None, None)
+        if isinstance(python_package_config, six.string_types)
+        else (
+            python_package_config['package_name'],
+            python_package_config.get('attribute'),
+            python_package_config.get('location_name'),
+        )
+    )
+
+
+def location_handle_from_package_name(package_name, attribute, location_name=None):
+    check.str_param(package_name, 'package_name')
+    check.opt_str_param(attribute, 'attribute')
+    check.opt_str_param(location_name, 'location_name')
+
+    loadable_targets = (
+        [LoadableTarget(attribute, load_def_in_package(package_name, attribute))]
+        if attribute
+        else loadable_targets_from_python_package(package_name)
+    )
+
+    repository_code_pointer_dict = {}
+    for loadable_target in loadable_targets:
+        repository_code_pointer_dict[
+            loadable_target.target_definition.name
+        ] = CodePointer.from_python_package(package_name, loadable_target.attribute)
+
+    return RepositoryLocationHandle.create_out_of_process_location(
+        repository_code_pointer_dict=repository_code_pointer_dict,
+        # default to the name of the repository symbol for now
+        location_name=assign_location_name(location_name, repository_code_pointer_dict),
+    )
+
+
 def assign_location_name(location_name, repository_code_pointer_dict):
     if location_name:
         return location_name
@@ -115,41 +161,53 @@ def assign_location_name(location_name, repository_code_pointer_dict):
 def location_handle_from_python_file_config(python_file_config, yaml_path):
     check.str_param(yaml_path, 'yaml_path')
 
-    absolute_path, attribute, location_name = _get_python_file_config_data(
+    absolute_path, attribute, location_name, working_directory = _get_python_file_config_data(
         python_file_config, yaml_path
     )
 
-    return location_handle_from_python_file(absolute_path, attribute, location_name)
+    return location_handle_from_python_file(
+        absolute_path, attribute, location_name, working_directory
+    )
 
 
 def _get_python_file_config_data(python_file_config, yaml_path):
     return (
-        (rebase_file(python_file_config, yaml_path), None, None)
+        (rebase_file(python_file_config, yaml_path), None, None, None)
         if isinstance(python_file_config, six.string_types)
         else (
             rebase_file(python_file_config['relative_path'], yaml_path),
             python_file_config.get('attribute'),
             python_file_config.get('location_name'),
+            rebase_file(python_file_config.get('working_directory'), yaml_path)
+            if python_file_config.get('working_directory')
+            else None,
         )
     )
 
 
-def location_handle_from_python_file(python_file, attribute, location_name=None):
+def location_handle_from_python_file(
+    python_file, attribute, location_name=None, working_directory=None
+):
     check.str_param(python_file, 'python_file')
     check.opt_str_param(attribute, 'attribute')
     check.opt_str_param(location_name, 'location_name')
+    check.opt_str_param(working_directory, 'working_directory')
 
     loadable_targets = (
-        [LoadableTarget(attribute, load_def_in_python_file(python_file, attribute))]
+        [
+            LoadableTarget(
+                attribute, load_def_in_python_file(python_file, attribute, working_directory)
+            )
+        ]
         if attribute
-        else loadable_targets_from_python_file(python_file)
+        else loadable_targets_from_python_file(python_file, working_directory)
     )
 
     repository_code_pointer_dict = {}
     for loadable_target in loadable_targets:
         repository_code_pointer_dict[
             loadable_target.target_definition.name
-        ] = CodePointer.from_python_file(python_file, loadable_target.attribute)
+        ] = CodePointer.from_python_file(python_file, loadable_target.attribute, working_directory)
 
     return RepositoryLocationHandle.create_out_of_process_location(
         repository_code_pointer_dict=repository_code_pointer_dict,
@@ -170,26 +228,32 @@ def _location_handle_from_python_environment_config(python_environment_config, y
 
     check.invariant(is_target_config(target_config))
 
-    python_file_config, python_module_config = (
+    python_file_config, python_module_config, python_package_config = (
         target_config.get('python_file'),
         target_config.get('python_module'),
+        target_config.get('python_package'),
     )
 
     if python_file_config:
-        absolute_path, attribute, location_name = _get_python_file_config_data(
+        absolute_path, attribute, location_name, working_directory = _get_python_file_config_data(
             python_file_config, yaml_path
         )
 
         if not attribute:
             response = sync_list_repositories(
-                executable_path=executable_path, python_file=absolute_path, module_name=None,
+                executable_path=executable_path,
+                python_file=absolute_path,
+                module_name=None,
+                working_directory=None,
             )
 
             return RepositoryLocationHandle.create_python_env_location(
                 executable_path=executable_path,
                 location_name=location_name,
                 repository_code_pointer_dict={
-                    lrs.attribute: CodePointer.from_python_file(absolute_path, lrs.attribute)
+                    lrs.attribute: CodePointer.from_python_file(
+                        absolute_path, lrs.attribute, working_directory
+                    )
                     for lrs in response.repository_symbols
                 },
             )
@@ -198,16 +262,21 @@ def _location_handle_from_python_environment_config(python_environment_config, y
                 executable_path=executable_path,
                 location_name=location_name,
                 repository_code_pointer_dict={
-                    attribute: CodePointer.from_python_file(absolute_path, attribute)
+                    attribute: CodePointer.from_python_file(
+                        absolute_path, attribute, working_directory
+                    )
                 },
             )
-    else:
+    elif python_module_config:
         check.invariant(python_module_config)
         module_name, attribute, location_name = _get_module_config_data(python_module_config)
 
         if not attribute:
             response = sync_list_repositories(
-                executable_path=executable_path, python_file=None, module_name=module_name,
+                executable_path=executable_path,
+                python_file=None,
+                module_name=module_name,
+                working_directory=None,
             )
             return RepositoryLocationHandle.create_python_env_location(
                 executable_path=executable_path,
@@ -223,6 +292,34 @@ def _location_handle_from_python_environment_config(python_environment_config, y
                 location_name=location_name,
                 repository_code_pointer_dict={
                     attribute: CodePointer.from_module(module_name, attribute)
+                },
+            )
+
+    else:
+        check.invariant(python_package_config)
+        package_name, attribute, location_name = _get_package_config_data(python_package_config)
+
+        if not attribute:
+            response = sync_list_repositories(
+                executable_path=executable_path,
+                python_file=None,
+                module_name=package_name,
+                working_directory=None,
+            )
+            return RepositoryLocationHandle.create_python_env_location(
+                executable_path=executable_path,
+                location_name=location_name,
+                repository_code_pointer_dict={
+                    lrs.attribute: CodePointer.from_python_package(package_name, lrs.attribute)
+                    for lrs in response.repository_symbols
+                },
+            )
+        else:
+            return RepositoryLocationHandle.create_python_env_location(
+                executable_path=executable_path,
+                location_name=location_name,
+                repository_code_pointer_dict={
+                    attribute: CodePointer.from_python_package(package_name, attribute)
                 },
             )
 
@@ -245,7 +342,9 @@ def _location_handle_from_location_config(location_config, yaml_path):
 
 def is_target_config(potential_target_config):
     return isinstance(potential_target_config, dict) and (
-        potential_target_config.get('python_file') or potential_target_config.get('python_module')
+        potential_target_config.get('python_file')
+        or potential_target_config.get('python_module')
+        or potential_target_config.get('python_package')
     )
 
 
@@ -259,5 +358,9 @@ def _location_handle_from_target_config(target_config, yaml_path):
 
     elif 'python_module' in target_config:
         return _location_handle_from_module_config(target_config['python_module'])
+
+    elif 'python_package' in target_config:
+        return _location_handle_from_package_config(target_config['python_package'])
+
     else:
         check.failed('invalid target_config')