+++ python_modules/dagster/dagster/core/definitions/reconstructable.py
@@ -6,6 +6,7 @@ from dagster import check, seven
 from dagster.core.code_pointer import (
     CodePointer,
     FileCodePointer,
+    FileInDirectoryCodePointer,
     ModuleCodePointer,
     get_python_file_from_previous_stack_frame,
 )
@@ -40,7 +41,9 @@ class ReconstructableRepository(namedtuple('_ReconstructableRepository', 'pointe
         return ReconstructableSchedule(self, name)
 
     @classmethod
-    def for_file(cls, file, fn_name):
+    def for_file(cls, file, fn_name, working_directory=None):
+        if working_directory:
+            return cls(FileInDirectoryCodePointer(file, fn_name, working_directory))
         return cls(FileCodePointer(file, fn_name))
 
     @classmethod