+++ python_modules/dagster/dagster_tests/cli_tests/workspace_tests/autodiscovery_tests/autodiscover_file_in_directory/autodiscover_src/pipelines.py
@@ -1,0 +1,11 @@
+from dagster import pipeline, solid
+
+
+@solid
+def hello_world(_):
+    pass
+
+
+@pipeline
+def hello_world_pipeline():
+    hello_world()