+++ python_modules/dagster/dagster_tests/cli_tests/workspace_tests/hello_world_file_in_directory/hello_world_repository.py
@@ -1,0 +1,8 @@
+from src.pipelines import hello_world_pipeline
+
+from dagster import repository
+
+
+@repository
+def hello_world_repository():
+    return [hello_world_pipeline]