+++ python_modules/dagster/dagster_tests/core_tests/test_instance.py
@@ -1,5 +1,3 @@
-import os
-
 import pytest
 
 from dagster import PipelineDefinition, execute_pipeline, pipeline, solid